Remove unused Android Studio configuration files
This commit removes `.idea/runConfigurations.xml` and updates `.idea/libraries/Flutter_Plugins.xml` as these files are specific to Android Studio and not essential for the project's functionality.

Additionally, `@pragma('vm:entry-point')` has been added to `callbackDispatcher` in `lib/src/plugin.dart` to mark it as an entry point for the Dart VM, which is relevant for background isolate execution.